packaging: adapt to lib/lib64 directory structure
1. Create aarch64 binary package from x86_64 build, since it's better
compatible with lib/lib64 directory structure. (in baselibs.conf)

2. Fixed path in symlinks in post-install script. (in baselibs.conf)

3. Disabled packaging of gconv locales that were packed for x86_64
build (that was not used before) from /usr/lib64/gconv/* .
This was done, because these locales are packed in glibc-locale
package for now (so these 2 packages conflict with each other.
(in qemu-accel-aarch64.spec)

4. Renamed BuildRequires tag for cross-aarch64-gcc package, since
we actually do not icecream build and the name was changef from the
side of gcc49 package. (in qemu-accel-aarch64.spec)

%{gcc_version}-icecream-backend

Change-Id: I4a98753c6e9229733162f4c922e42e3aa5268549
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>